+++ src/main/java/calculator/domain/Expression.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 public interface Expression {
     String[] array();
