+++ src/main/java/calculator/domain/StringAdditionExpression.java
@@ -1,10 +1,10 @@
-package calculator;
+package calculator.domain;
 
 import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class StringAdditionExpression implements Expression{
+public class StringAdditionExpression implements Expression {
     public static final String DEFAULT = "[,:]";
     public static final String CUSTOM_PREFIX = "//";
     public static final String CUSTOM_SUFFIX = "\n";
