+++ src/test/java/calculator/domain/CalculatorTest.java
@@ -1,5 +1,8 @@
-package calculator;
+package calculator.domain;
 
+import calculator.domain.Calculator;
+import calculator.domain.Expression;
+import calculator.domain.StringAdditionExpression;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
