+++ src/test/java/calculator/domain/StringAdditionExpressionTest.java
@@ -1,5 +1,6 @@
-package calculator;
+package calculator.domain;
 
+import calculator.domain.StringAdditionExpression;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
